+++ src/aiheuristics/EmptySpaces.java
@@ -32,11 +32,11 @@ public class EmptySpaces extends Heuristic {
     
     @Override
     public long getValueOfState(GameController controller, GameBoard state, int currentDirection) {
-        byte[][] grid = state.getGameGrid();
+        int[][] grid = state.getGameGrid();
         int[] highestPos = getHighestValuePosition(grid);
         int highestValue = grid[highestPos[0]][highestPos[1]];
         long output = 0;
-        for (byte[] row : grid) {
+        for (int[] row : grid) {
             for (int value : row) {
                 if (value == 0) {
                     output += (highestValue / 2);
@@ -48,11 +48,11 @@ public class EmptySpaces extends Heuristic {
     }
 
 // returned as y, x
-    private int[] getHighestValuePosition(byte[][] grid) {
+    private int[] getHighestValuePosition(int[][] grid) {
         int[] highest = new int[2];
         int highestValue = 0;
         for (int y = 0; y < grid.length; y++) {
-            byte[] row = grid[y];
+            int[] row = grid[y];
             for (int x = 0; x < row.length; x++) {
                 int value = row[x];
                 if (value > highestValue) {
